+++ SurfaceEvolver/Geometry/Matrix3.h
@@ -6,35 +6,35 @@
 class Matrix3
 {
 public:
-	float elements[9] = {
-		1.0f, 0.0f, 0.0f,
-		0.0f, 1.0f, 0.0f,
-		0.0f, 0.0f, 1.0f
+	double elements[9] = {
+		1.0, 0.0, 0.0,
+		0.0, 1.0, 0.0,
+		0.0, 0.0, 1.0
 	};
 
 	Matrix3();
 	Matrix3(const Matrix3& other);
-	Matrix3(float* elems);
+	Matrix3(double* elems);
 	~Matrix3();
 
 	bool isIdentity();
 	void setToIdentity();
 	void set(
-		float m00, float m01, float m02,
-		float m10, float m11, float m12,
-		float m20, float m21, float m22
+		double m00, double m01, double m02,
+		double m10, double m11, double m12,
+		double m20, double m21, double m22
 	);
 
 	void transpose();
-	void multiplyScalar(float scalar);
+	void multiplyScalar(double scalar);
 
-	float determinant();
+	double determinant();
 	Matrix3 getInverse(Matrix3& from);
 
 	Matrix3 operator+ (Matrix3 other);
 	Matrix3 operator- (Matrix3 other);
 	Matrix3 operator* (Matrix3 other);
-	friend Matrix3 operator*(float scalar, Matrix3 m);
+	friend Matrix3 operator*(double scalar, Matrix3 m);
 };
 
 Matrix3 inverse(Matrix3& m);
